Refactor: add Marked and SeqMarked to replace TSeqValue
The structure of `TSeqValue` does not make sense:
A `TSeqValue` tombstone does not need to store user data.

We replace `TSeqValue` with `Marked` for storing a tombstone data and
`SeqMarked` that binds seq number to a `Marked` data.

`SeqMarked` also implements a natural partially order, which considers
tombstone data greater than normal data.